meson.build: clean up ffmsindex dependency on ffms2
Previously, ffms2 was required as a dependency, leading to it
needing to have been installed to the system, and failing if it
wasn't.

Instead of doing that, meson provides a mechanism to re-use objects
from an already built library in the same build process, which is
essentially what ffmsindex linking to ffms2 was doing anyway.